move testCase enum
