+++ examples/src/visual_tests/basic.test.ts
@@ -1,20 +1,10 @@
+import { TestCase } from './TestCase';
+
 const retry = require('jest-retries');
 
 const RETRIES = parseInt(process.env.RETRIES!) || 3;
 
 jest.retryTimes(RETRIES);
-
-export enum TestCase {
-  default = 'default',
-  clipping = 'clipping',
-  default_camera = 'default_camera',
-  highlight = 'highlight',
-  rotate_cad_model = 'rotate_cad_model',
-  node_transform = 'node_transform',
-  ghost_mode = 'ghost_mode',
-  scaled_model = 'scaled_model',
-  user_render_target = 'user_render_target',
-}
 
 async function screenShotTest(testCase: TestCase) {
   const page = await browser.newPage();
